requirements2.txt
absl-py==1.2.0
anyio==3.6.1
asttokens==2.0.8
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
executing==0.10.0
fastapi==0.79.0
fonttools==4.35.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646982648551/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///opt/conda/conda-bld/ipython_1657652213665/work
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1650622202839/work
jupyter-core @ file:///opt/conda/conda-bld/jupyter_core_1651671229925/work
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847908682/work
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-multipart==0.0.5
pytz==2022.2.1
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
stack-data==0.4.0
starlette==0.19.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
uvicorn==0.18.2
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
Werkzeug==2.2.2
zipp==3.8.1
